qqplotr {qqplotr}	R Documentation

Q-Q and P-P plot extensions for 'ggplot2'

Description

This package extends some ggplot2 functionalities by permitting the drawing of both quantile-quantile (Q-Q) and probability-probability (P-P) points, lines, and confidence bands. The functions of this package also allow the detrend adjustment, proposed by Thode (2002), which helps reduce visual bias when assessing those plots.

Details

The functions of this package, presented as ggplot2 Stats, are divided into two groups: Q-Q and P-P related.

Each of the groups is composed of three Stats: point, line, and band. Those Stats, while independent, complement each other when plotted together.

[Package qqplotr version 0.0.5 Index]